Update password and profile of currently login usr
In this PR, we update the password and profile of currently login user. We use the middleware to check if the user is login or not. If the user is login, we will update the password and profile of the user. If the user is not login, we will return the error message.